+++ GarageManagement.Server/Model/JobCardConcern.cs
@@ -2,8 +2,8 @@
 {
     public class JobCardConcern
     {
-        public int Id { get; set; }
-        public int JobCardId { get; set; }
+        public long Id { get; set; }
+        public long JobCardId { get; set; }
 
         public string? Text { get; set; }
         public bool Active { get; set; }
